+++ src/Akeneo/Pim/Enrichment/Component/Product/Query/GetProductUuidsWithRemovedAttributeInterface.php
@@ -4,7 +4,7 @@ namespace Akeneo\Pim\Enrichment\Component\Product\Query;
 
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\SearchQueryBuilder;
 
-interface GetProductIdentifiersWithRemovedAttributeInterface
+interface GetProductUuidsWithRemovedAttributeInterface
 {
     public function nextBatch(array $attributesCodes, int $batchSize): iterable;