+++ tests/back/Pim/Enrichment/Integration/Product/RemoveValuesFromProductsAndProductModelsIntegration.php
@@ -7,6 +7,7 @@ namespace AkeneoTest\Pim\Enrichment\Integration\Product;
 use Akeneo\Pim\Enrichment\Bundle\Product\RemoveValuesFromProductModels;
 use Akeneo\Pim\Enrichment\Bundle\Product\RemoveValuesFromProducts;
 use Akeneo\Pim\Enrichment\Component\Product\Message\ProductUpdated;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Structure\Component\AttributeTypes;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Test\Common\EntityBuilder;
@@ -15,6 +16,7 @@ use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
 use Akeneo\Test\IntegrationTestsBundle\Messenger\AssertEventCountTrait;
 use Doctrine\DBAL\Connection;
+use Ramsey\Uuid\Uuid;
 
 class RemoveValuesFromProductsAndProductModelsIntegration extends TestCase
 {
@@ -28,10 +30,10 @@ class RemoveValuesFromProductsAndProductModelsIntegration extends TestCase
         $niceAttribute = $this->createAttribute('nice_attribute');
         $otherAttribute = $this->createAttribute('other_attribute');
 
-        $this->createProductWithAttributeValue('product_one', 'nice_attribute', 'value one');
-        $this->createProductWithAttributeValue('product_two', 'nice_attribute', 'value two');
-        $this->createProductWithAttributeValue('product_three', 'other_attribute', 'value three');
-        $this->createProductWithAttributeValue('product_four', 'other_attribute', 'value four');
+        $product1 = $this->createProductWithAttributeValue('product_one', 'nice_attribute', 'value one');
+        $product2 = $this->createProductWithAttributeValue('product_two', 'nice_attribute', 'value two');
+        $product3 = $this->createProductWithAttributeValue('product_three', 'other_attribute', 'value three');
+        $product4 = $this->createProductWithAttributeValue('product_four', 'other_attribute', 'value four');
 
         $this->deleteAttribute($niceAttribute);
         $this->deleteAttribute($otherAttribute);
@@ -40,19 +42,19 @@ class RemoveValuesFromProductsAndProductModelsIntegration extends TestCase
             'nice_attribute',
             'other_attribute'
         ];
-        $productIdentifiers = [
-            'product_one',
-            'product_two',
-            'product_three',
-            'product_four'
+        $productUuids = [
+            $product1->getUuid()->toString(),
+            $product2->getUuid()->toString(),
+            $product3->getUuid()->toString(),
+            $product4->getUuid()->toString(),
         ];
 
-        $this->assertEquals(4, $this->getProductWithAttributeValuesCount($attributeCodes, $productIdentifiers));
+        $this->assertEquals(4, $this->getProductWithAttributeValuesCount($attributeCodes, $productUuids));
 
         $this->getRemoveValuesFromProducts()
-            ->forAttributeCodes($attributeCodes, $productIdentifiers);
+            ->forAttributeCodes($attributeCodes, $productUuids);
 
-        $this->assertEquals(0, $this->getProductWithAttributeValuesCount($attributeCodes, $productIdentifiers));
+        $this->assertEquals(0, $this->getProductWithAttributeValuesCount($attributeCodes, $productUuids));
 
         $this->assertEventCount(4, ProductUpdated::class);
     }
@@ -92,10 +94,11 @@ class RemoveValuesFromProductsAndProductModelsIntegration extends TestCase
         $this->assertEquals(0, $this->getProductModelWithAttributeValuesCount($attributeCodes, $productModelCodes));
     }
 
-    private function getProductWithAttributeValuesCount(array $attributeCodes, array $productIdentifiers): int
+    private function getProductWithAttributeValuesCount(array $attributeCodes, array $productUuids): int
     {
         $connection = $this->get('database_connection');
 
+        $uuidsAsBytes = \array_map(fn($productUuid) => Uuid::fromString($productUuid)->getBytes(), $productUuids);
         $paths = implode(
             ',',
             array_map(fn ($attributeCode) => $connection->quote(sprintf('$."%s"', $attributeCode)), $attributeCodes)
@@ -106,14 +109,14 @@ class RemoveValuesFromProductsAndProductModelsIntegration extends TestCase
 SELECT COUNT(id)
 FROM `pim_catalog_product`
 WHERE JSON_CONTAINS_PATH(raw_values, 'one', $paths)
-AND identifier IN (:product_identifiers)
+AND uuid IN (:product_uuids)
 SQL,
             [
-                'product_identifiers' => $productIdentifiers
+                'product_uuids' => $uuidsAsBytes
             ],
             [
 
-                'product_identifiers' => Connection::PARAM_STR_ARRAY
+                'product_uuids' => Connection::PARAM_STR_ARRAY
             ]
         )->fetchOne();
 
@@ -180,11 +183,13 @@ SQL,
         string $identifier,
         string $attributeCode,
         string $valueData
-    ): void {
+    ): ProductInterface {
         $product = $this->getProductBuilder()
             ->withIdentifier($identifier)
             ->withValue($attributeCode, $valueData)->build();
         $this->get('pim_catalog.saver.product')->save($product);
+
+        return $product;
     }
 
     private function createProductModelWithAttributeValue(